+++ src/main/java/project/springmarket/mapper/MemberMapper.java
@@ -1,4 +1,4 @@
-package project.springmarket.mybatis;
+package project.springmarket.mapper;
 
 import org.apache.ibatis.annotations.Mapper;
 import project.springmarket.model.member.Member;
