+++ src/main/java/project/springmarket/repository/MemberRepository.java
@@ -3,7 +3,7 @@ package project.springmarket.repository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import project.springmarket.model.member.Member;
-import project.springmarket.mybatis.MemberMapper;
+import project.springmarket.mapper.MemberMapper;
 
 @RequiredArgsConstructor
 @Repository
